+++ app/mailers/user_mailer.rb
@@ -10,7 +10,6 @@ class UserMailer < ActionMailer::Base
 
   default from: Rails.configuration.x.organisation.email
 
-  # rubocop:disable Metrics/AbcSize
   def welcome_notification(user)
     @user           = user
     @username       = @user.name
@@ -20,12 +19,9 @@ class UserMailer < ActionMailer::Base
     @contact_us     = Rails.application.config.x.organisation.contact_us_url || contact_us_url
     @helpdesk_email = helpdesk_email(org: @user.org)
 
-    I18n.with_locale I18n.locale do
-      mail(to: @user.email,
-           subject: format(_('Welcome to %{tool_name}'), tool_name: tool_name))
-    end
+    mail(to: @user.email,
+         subject: "Welcome to DMP Assistant / Bienvenue sur l'Assistant PGD")
   end
-  # rubocop:enable Metrics/AbcSize
 
   # rubocop:disable Metrics/AbcSize
   def question_answered(data, user, answer, _options_string)
@@ -41,7 +37,9 @@ class UserMailer < ActionMailer::Base
     @answer_text    = @options_string.to_s
     @helpdesk_email = helpdesk_email(org: @user.org)
 
-    I18n.with_locale I18n.locale do
+    recipient = User.find_by(email: data['email'])
+
+    LocaleService.with_preferred_locale(recipient) do
       mail(to: data['email'],
            subject: data['subject'])
     end
@@ -52,12 +50,14 @@ class UserMailer < ActionMailer::Base
     @role       = role
     @user       = user
     @user_email = @user.email
+    # TODO: Should this be `@user.name(false)`? Currently, @username is just being assigned @user.email
+    # (There are several other user.name calls in UserMailer that should be verified as well)
     @username   = @user.name
     @inviter    = inviter
     @link       = url_for(action: 'show', controller: 'plans', id: @role.plan.id)
     @helpdesk_email = helpdesk_email(org: @inviter.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(@role.user) do
       mail(to: @role.user.email,
            subject: format(_('A Data Management Plan in %{tool_name} has been shared with you'),
                            tool_name: tool_name))
@@ -74,7 +74,7 @@ class UserMailer < ActionMailer::Base
     @messaging = role_text(@role)
     @helpdesk_email = helpdesk_email(org: @user.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(@recepient) do
       mail(to: @recepient.email,
            subject: format(_('Changed permissions on a Data Management Plan in %{tool_name}'),
                            tool_name: tool_name))
@@ -89,7 +89,7 @@ class UserMailer < ActionMailer::Base
     @current_user = current_user
     @helpdesk_email = helpdesk_email(org: @plan.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(@user) do
       mail(to: @user.email,
            subject: format(_('Permissions removed on a DMP in %{tool_name}'),
                            tool_name: tool_name))
@@ -107,7 +107,7 @@ class UserMailer < ActionMailer::Base
     @plan_name      = @plan.title
     @helpdesk_email = helpdesk_email(org: @plan.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(@recipient) do
       mail(to: @recipient.email,
            subject: format(_('%{user_name} has requested feedback on a %{tool_name} plan'),
                            tool_name: tool_name, user_name: @user.name(false)))
@@ -126,7 +126,7 @@ class UserMailer < ActionMailer::Base
     @plan_name      = @plan.title
     @helpdesk_email = helpdesk_email(org: @plan.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(recipient) do
       sender = Rails.configuration.x.organisation.do_not_reply_email ||
                Rails.configuration.x.organisation.email
 
@@ -148,7 +148,7 @@ class UserMailer < ActionMailer::Base
     @plan_visibility = Plan::VISIBILITY_MESSAGE[@plan.visibility.to_sym]
     @helpdesk_email = helpdesk_email(org: @plan.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(@user) do
       mail(to: @user.email,
            subject: format(_('DMP Visibility Changed: %{plan_title}'), plan_title: @plan.title))
     end
@@ -177,7 +177,7 @@ class UserMailer < ActionMailer::Base
     @phase_link = url_for(action: 'edit', controller: 'plans', id: @plan.id, phase_id: @phase_id)
     @helpdesk_email = helpdesk_email(org: @plan.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(@plan.owner) do
       mail(to: @plan.owner.email,
            subject: format(_('%{tool_name}: A new comment was added to %{plan_title}'),
                            tool_name: tool_name, plan_title: @plan.title))
@@ -190,10 +190,9 @@ class UserMailer < ActionMailer::Base
 
     @user      = user
     @username  = @user.name
-    @ul_list   = privileges_list(@user)
     @helpdesk_email = helpdesk_email(org: @user.org)
 
-    I18n.with_locale I18n.locale do
+    LocaleService.with_preferred_locale(@user) do
       mail(to: user.email,
            subject: format(_('Administrator privileges updated in %{tool_name}'),
                            tool_name: tool_name))
@@ -211,7 +210,9 @@ class UserMailer < ActionMailer::Base
 
     @helpdesk_email = helpdesk_email(org: @api_client.org)
 
-    I18n.with_locale I18n.locale do
+    recipient = User.find_by(email: @api_client.contact_email) || @api_client.org
+
+    LocaleService.with_preferred_locale(recipient) do
       mail(to: @api_client.contact_email,
            subject: format(_('%{tool_name} API changes'), tool_name: tool_name))
     end